Fix GTK4 Linux build errors and GCC narrowing warnings
Replace removed GTK2 main-loop helpers in lv2gui.cpp:
gtk_events_pending/gtk_main_iteration -> g_main_context_pending/
g_main_context_iteration; remove redundant gtk_widget_show/hide calls
already superseded by gtk_window_present/gtk_window_destroy.

Fix -Wnarrowing in GRAPHENE_RECT_INIT across 14 widget files: GCC
treats int-to-float as an error in aggregate initializers, requiring
explicit (float) casts on width/height arguments.

+++ src/ctl_buttons.cpp
@@ -73,7 +73,7 @@ calf_toggle_snapshot (GtkWidget *widget, GtkSnapshot *snapshot)
     float x = wcx - pcx;
     float y = wcy - pcy;
 
-    graphene_rect_t bounds = GRAPHENE_RECT_INIT(0, 0, gtk_widget_get_width(widget), gtk_widget_get_height(widget));
+    graphene_rect_t bounds = GRAPHENE_RECT_INIT(0, 0, (float)gtk_widget_get_width(widget), (float)gtk_widget_get_height(widget));
     cairo_t *c = gtk_snapshot_append_cairo(snapshot, &bounds);
     calf_cairo_set_source_pixbuf(c, self->toggle_image, x - 0, y - sy);
     cairo_rectangle(c, x, y, pw, ph / 2);
@@ -253,7 +253,7 @@ calf_button_snapshot (GtkWidget *widget, GtkSnapshot *snapshot)
         int width  = gtk_widget_get_width(widget);
         int height = gtk_widget_get_height(widget);
 
-        graphene_rect_t bounds = GRAPHENE_RECT_INIT(0, 0, gtk_widget_get_width(widget), gtk_widget_get_height(widget));
+        graphene_rect_t bounds = GRAPHENE_RECT_INIT(0, 0, (float)gtk_widget_get_width(widget), (float)gtk_widget_get_height(widget));
         cairo_t *c = gtk_snapshot_append_cairo(snapshot, &bounds);
 
         int x  = 0;
@@ -514,7 +514,7 @@ calf_tap_button_snapshot (GtkWidget *widget, GtkSnapshot *snapshot)
     int x = 0 + width / 2 - img_width / 2;
     int y = 0 + height / 2 - img_height / 2;
 
-    graphene_rect_t bounds = GRAPHENE_RECT_INIT(0, 0, gtk_widget_get_width(widget), gtk_widget_get_height(widget));
+    graphene_rect_t bounds = GRAPHENE_RECT_INIT(0, 0, (float)gtk_widget_get_width(widget), (float)gtk_widget_get_height(widget));
     cairo_t *c = gtk_snapshot_append_cairo(snapshot, &bounds);
     calf_cairo_set_source_pixbuf(c, self->image[self->state], x - 0, y - 0);
     cairo_rectangle(c, x, y, img_width, img_height);
