+++ src/ctl_fader.cpp
@@ -217,7 +217,7 @@ calf_fader_snapshot (GtkWidget *widget, GtkSnapshot *snapshot)
 
         int width  = gtk_widget_get_width(widget);
         int height = gtk_widget_get_height(widget);
-        graphene_rect_t bounds = GRAPHENE_RECT_INIT(0, 0, width, height);
+        graphene_rect_t bounds = GRAPHENE_RECT_INIT(0, 0, (float)width, (float)height);
         cairo_t *c = gtk_snapshot_append_cairo(snapshot, &bounds);
 
         cairo_rectangle(c, l.x, l.y, l.w, l.h);
